refactor: extracted GenericValue into generic_value.py
