+++ src/inline_snapshot/_inline_snapshot.py
@@ -19,116 +19,15 @@ from ._change import Replace
 from ._code_repr import code_repr
 from ._exceptions import UsageError
 from ._sentinels import undefined
-from ._types import Snapshot
+from ._snapshot.generic_value import GenericValue
 from ._unmanaged import map_unmanaged
 from ._unmanaged import Unmanaged
-from ._unmanaged import update_allowed
 from ._utils import value_to_token
 from .global_state import state
 
 
-def _return(result):
-    if not result:
-        state()._incorrect_values += 1
-    return result
-
-
 def ignore_old_value():
     return state()._update_flags.fix or state()._update_flags.update
-
-
-class GenericValue(Snapshot):
-    _new_value: Any
-    _old_value: Any
-    _current_op = "undefined"
-    _ast_node: ast.Expr
-    _context: AdapterContext
-
-    @property
-    def _file(self):
-        return self._context.file
-
-    def get_adapter(self, value):
-        return get_adapter_type(value)(self._context)
-
-    def _re_eval(self, value, context: AdapterContext):
-        self._context = context
-
-        def re_eval(old_value, node, value):
-            if isinstance(old_value, Unmanaged):
-                old_value.value = value
-                return
-
-            assert type(old_value) is type(value)
-
-            adapter = self.get_adapter(old_value)
-            if adapter is not None and hasattr(adapter, "items"):
-                old_items = adapter.items(old_value, node)
-                new_items = adapter.items(value, node)
-                assert len(old_items) == len(new_items)
-
-                for old_item, new_item in zip(old_items, new_items):
-                    re_eval(old_item.value, old_item.node, new_item.value)
-
-            else:
-                if update_allowed(old_value):
-                    if not old_value == value:
-                        raise UsageError(
-                            "snapshot value should not change. Use Is(...) for dynamic snapshot parts."
-                        )
-                else:
-                    assert False, "old_value should be converted to Unmanaged"
-
-        re_eval(self._old_value, self._ast_node, value)
-
-    def _ignore_old(self):
-        return (
-            state()._update_flags.fix
-            or state()._update_flags.update
-            or state()._update_flags.create
-            or self._old_value is undefined
-        )
-
-    def _visible_value(self):
-        if self._ignore_old():
-            return self._new_value
-        else:
-            return self._old_value
-
-    def _get_changes(self) -> Iterator[Change]:
-        raise NotImplementedError()
-
-    def _new_code(self):
-        raise NotImplementedError()
-
-    def __repr__(self):
-        return repr(self._visible_value())
-
-    def _type_error(self, op):
-        __tracebackhide__ = True
-        raise TypeError(
-            f"This snapshot cannot be use with `{op}`, because it was previously used with `{self._current_op}`"
-        )
-
-    def __eq__(self, _other):
-        __tracebackhide__ = True
-        self._type_error("==")
-
-    def __le__(self, _other):
-        __tracebackhide__ = True
-        self._type_error("<=")
-
-    def __ge__(self, _other):
-        __tracebackhide__ = True
-        self._type_error(">=")
-
-    def __contains__(self, _other):
-        __tracebackhide__ = True
-        self._type_error("in")
-
-    def __getitem__(self, _item):
-        __tracebackhide__ = True
-        self._type_error("snapshot[key]")
 
 
 class UndecidedValue(GenericValue):
