+++ src/inline_snapshot/_snapshot/collection_value.py
@@ -5,7 +5,6 @@ from .._change import Change
 from .._change import Delete
 from .._change import ListInsert
 from .._change import Replace
-from .._inline_snapshot import _return
 from .._inline_snapshot import clone
 from .._inline_snapshot import GenericValue
 from .._inline_snapshot import ignore_old_value
@@ -30,7 +29,7 @@ class CollectionValue(GenericValue):
         if ignore_old_value() or self._old_value is undefined:
             return True
         else:
-            return _return(item in self._old_value)
+            return self._return(item in self._old_value)
 
     def _new_code(self):
         return self._file._value_to_code(self._new_value)
